Totally reconfigured and rearanged the way the pdfCreater.java is setup and constructed. While there are still much more work to be done in terms of easy overview and readability, this first major clean up surely makes it a lot more easier to jump in than previously.
MAJOR overhaul of the pdfCreater.java!
The code has been cut up into a lot more methods, making is WAY MORE EASIER
to go through, understand and get a good overview.

Also changed the way a lot pictures are handled, making a lot of pictures 
obsolete!

Needs a lot of commenting still though + PDFont

Added a new method for checkBoxes on Page 2.

Added a couple of new methods to Page3.