Added a few more fields og text and input boxes for furture use in the createPDF.jsp.
MAJOR Revision of the PDFCreator.java!
Added A LOT of comments!
Pages can now be (very basic) be auto generated, based on the amount of pages that you want and all are created with the required basic layout. 
The old 'hardcoded'page numbers are now dynamic, to fit the autogenerated and possible infinite amount of pages.
Also done some major work on the 'Gennemgang af lokale', in order to fit
the dynamic creation of pages. This is not totally complete, but its getting 
there.